\name{Internal functions}
\alias{choicefactGUI}
\alias{choiceGUI}
\alias{cluster_det_URF}
\alias{correlation.RF.init.vs.fitted}
\alias{Davies_Bouldin_URF}
\alias{dev_delete}
\alias{diversity.calc.RF}
\alias{fig_par}
\alias{graphfact.RF}
\alias{heat.mapGUI.RF}
\alias{import_variable.RF}
\alias{language.RF}
\alias{legend_color_scale}
\alias{loadGUI.RF}
\alias{MDA.RF}
\alias{mds.fitted.RF}
\alias{mds.RF}
\alias{mds.URF}
\alias{memory.increase}
\alias{misclassification_ini}
\alias{new.SRF.OOB.aff}
\alias{new.SRF.OOB}
\alias{OOB.RF.fitted}
\alias{OOB.RF}
\alias{perm.MDA.init}
\alias{plot_adjust.numb.var}
\alias{plot_univ}
\alias{RISA2RF.RF}
\alias{saveGUI.RF}
\alias{several_URF}
\alias{selection.RF}
\alias{SRF.fin.tune.mtryGUI}
\alias{SRF.tune.mtryGUI}
\alias{SRF_init_rpart}
\alias{SRF_fitted_rpart}
\alias{SRF_init_VarImp}
\alias{supRFGUI.fitted.RF}
\alias{supRFGUI.RF}
\alias{tk_select}
\alias{unsup.mode.RF}
\alias{URF_tune_mtry}
\alias{waitGUI}
\title{Internal functions for the RandForestGUI package}
\description{
Internal functions for the RandForestGUI package
}
\details{
These are not to be called by the user
}

\author{Rory Michelland \email{rory.michelland@gmail.com}\cr}
\seealso{\code{\link{RandForestGUI}}}
\keyword{microarray,sequence,fingerprint,random forest,multivariate,conditional inference}